+++ src/main/kotlin/de/sharetopia/productservice/product/service/RentRequestServiceImpl.kt
@@ -1,14 +1,13 @@
 package de.sharetopia.productservice.product.service
 
-import de.sharetopia.productservice.product.exception.ProductNotFoundException
+import de.sharetopia.productservice.product.exception.NotAllowedAccessToResourceException
+import de.sharetopia.productservice.product.exception.RentRequestNotFoundException
 import de.sharetopia.productservice.product.model.RentRequestModel
-import de.sharetopia.productservice.product.repository.ProductRepository
 import de.sharetopia.productservice.product.repository.RentRequestRepository
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import java.util.*
 
 @Service
 class RentRequestServiceImpl : RentRequestService {
@@ -16,7 +15,7 @@ class RentRequestServiceImpl : RentRequestService {
     private lateinit var rentRequestRepository: RentRequestRepository
 
     @Autowired
-    private lateinit var productRepository: ProductRepository
+    private lateinit var productService: ProductService
 
     private val log: Logger = LoggerFactory.getLogger(RentRequestServiceImpl::class.java)
 
@@ -24,10 +23,7 @@ class RentRequestServiceImpl : RentRequestService {
 
     override fun create(rentRequest: RentRequestModel, userId: String): RentRequestModel {
         rentRequest.requesterUserId = userId
-        val requestedProduct = productRepository.findById(rentRequest.requestedProductId!!).orElseThrow {
-            log.error("Error fetching requested product by id. {error=ProductNotFoundException, method=POST, endpoint=/products, requesterUserId=${userId}}")
-            ProductNotFoundException(rentRequest.requestedProductId!!)
-        }
+        val requestedProduct = productService.findById(rentRequest.requestedProductId!!)
         rentRequest.rentRequestReceiverUserId = requestedProduct.ownerOfProductUserId
         return rentRequestRepository.save(rentRequest)
     }
@@ -37,11 +33,20 @@ class RentRequestServiceImpl : RentRequestService {
         return rentRequestRepository.save(rentRequest)
     }
 
-    override fun findById(rentRequestId: String): Optional<RentRequestModel> {
-        return rentRequestRepository.findById(rentRequestId)
+    override fun findById(rentRequestId: String): RentRequestModel {
+        val rentRequest = rentRequestRepository.findById(rentRequestId).orElseThrow {
+            log.error("Error fetching rent request by id. {error=RentRequestNotFoundException, rentRequestId=$rentRequestId")
+            RentRequestNotFoundException(rentRequestId)
+        }
+        return rentRequest
     }
 
-    override fun deleteById(rentRequestId: String) {
+    override fun deleteById(rentRequestId: String, userId: String) {
+        val rentRequestToBeDeleted = findById(rentRequestId)
+        if (rentRequestToBeDeleted.requesterUserId != userId) {
+            log.error("Error by not allowed access to rent request. {NotAllowedAccessToResourceException, rentRequestId=$rentRequestId")
+            throw NotAllowedAccessToResourceException(userId)
+        }
         rentRequestRepository.deleteById(rentRequestId)
     }
